Struct farbfeld::FarbfeldDecoder

pub struct FarbfeldDecoder<R> {
    // some fields omitted
}

A farbfeld decoder

Methods

impl<R: Read + Seek> FarbfeldDecoder<R>

fn new(r: R) -> FarbfeldResult<FarbfeldDecoder<R>>

Create a new decoder from r and parse the header.

Failures

Returns FarbfeldError::FormatError if the magic number does not match farbfeld

fn dimensions(&mut self) -> (u32, u32)

Returns the (width, height) of the image.

fn row_len(&mut self) -> usize

Returns the length in bytes for a row.

fn read_row(&mut self, row: u32, buf: &mut [u8]) -> FarbfeldResult<usize>

Read a single row from the image and return the bytes read.

Failures

Returns a FarbfeldError::ImageEnd if the row is greater as the height

fn read_image(&mut self) -> FarbfeldResult<Vec<u8>>

Read whole image into a Vec<u8>.