Struct farbfeld::FarbfeldEncoder

pub struct FarbfeldEncoder<W: Write>(pub W);

A farbfeld encoder

Methods

impl<W: Write> FarbfeldEncoder<W>

fn encode(self, width: u32, height: u32, data: &[u8]) -> FarbfeldResult<()>

Encodes a image with width, height and data into a farbfeld.

Failures

Returns a FarbfeldError::NotEnoughData if the provided data slice is too short.